[tests] Generate dummy header files for GCC tests.
As a part of the cleanup, the '*_dummy.h' files in `tests/include` were removed,
as these needed to generated. The headers are generated in the `Makefile` but
not in `validate.sh`. This patch fixes that.
We eventually intend to move to using the `Makefile` for building and testing,
until then we need to support building from both `Makefile` and `validate.sh`.