[compiler] Save processed ACFILES relative to CWD.
(`CWD` -> Current Working Directory.)

The compiler originally created the processed `ACFILES` in the directory
that they were. It was not an issue earlier, since we had links to
libgum locally. Now, with `libgum` moved to a standard directory, we need
to make sure that the generated files do not pollute the standard `libgum`
directory and are saved in the `dist-dir` instead, and this tiny patch does
exactly that.