[skelfs] Skel-fs framework.
Add entry and exit functions to the various `super_block`, `file_operations`
and `inode_operations` structures.
The idea is to provide a framework of sorts that can be used to implement an
actual filesystem, by implementing the missing pieces, with the grunt work
already done.

This commit also adds support for verbose mode for Makefiles.