[libgum] Some methods to access struct buffer_head (#78)
* [libgum] Some methods to access struct buffer_head

We currently have `OSBuffer`(in osbuffer.cogent and osbuffer.ac) which is
a wrapper over `struct buffer_head`. This patch creates a new
implementation called `BufferHead`, which reflects the name of the actual
structure in the Linux kernel. And also cleans up the current
implementation.
I've created new files for this implementation (`bufferhead.cogent` and
`bufferhead.ac`) and retaining the old implementation until we transition
over to the new implementation everywhere.

This patch also adds:
* `SysState` to `common.cogent`, which is the eventual replacement for
  `ExState`. We are still retaining `ExState` until we make sure that
  transition has occured smoothly.
* Created a common `abstract-defns.h` file in `lib/c/linux` - this file
  is going to hold the definitions of all the abstract data types that
  the filesystem implementations are going to use.