[impl/fs] Fix ext2 and bilby build issues.
We don't want to build the kernel module(yet) since travis container doesn't
have the kernel headers that are required to build the kernel module. So what
we do is just check if the C code is generated successfully by the Cogent
compiler.

We also encountered a really interesting issue as highlighted in this[1]
stackoverflow post. So we run the initial pass of the `cpp` (from within Cogent)
in the `c99` mode and then when actually building the kernel module we use `cpp`
in `gnu99` mode. Please read the stackoverflow post for more details.

[1] - http://stackoverflow.com/questions/19210935/why-does-the-c-preprocessor-interpret-the-word-linux-as-the-constant-1