[impl] Port fs modules to kernel version > 4.4
This patch ports the Cogent fs modules `vfat` and `ext2` to support the
recent kernels(4.4 and later).

This patch doesn't include port the bilby implementation, since it isn't
as straightforward. I'm working on that patch, and will send it soon.

I've also removed most code to run on kernel version < 3.16, we don't support
them anymore.

Fixes #65.