compiler: proof gen hygiene: explicitly quantify variables where possible
Some of the generated lemmas and definitions were using names like 'x',
which failed when there was a Cogent function called 'x', because they
end up referring to the Cogent function instead of defining a new variable.

We need a better story about this in the future, but for now I've just
tried to add quantifiers where I can.