compiler: shallow: Case: don't use undefined for impossible cases
the specialised Case rules for the shallow correspondence proof require
all branches in the shallow case to be the same, except the matched tag.
this means we cannot generate cases with different branches. if we know
that a particular variant constructor is not possible, we should be able
to reuse the rest of the continuation, without any problems.