compiler: shallow embedding & correspondence: let-bind case continuation
In the shallow embedding for case expressions, we were generating
shallow embeddings as the following:
```
Case tag scrut when_tag when_other
==>
(\continuation.
  case_T scrut (\x. when_tag x)
               (\x. continuation (OtherTag1 x))
               (\x. continuation (OtherTag2 x)))
  (\y. when_other y)
```

However, because Isabelle performs eager beta reduction, this ends up
duplicating `when_other` into both other tags. If it's a variant with a
lot of constructors, or a large continuation, then this duplication can
cause a pretty serious explosion in term size.

This commit changes code generation to use a let binding (HOL.Let), and
updates the `scorres_case_T__tag` specialised lemmas to work on the let
binding.

This change might affect the top-level Bilby proofs, if they are relying
on the continuation of a case to be simplified away. If so, it is
hopefully sufficient to just add something like `(simp only: HOL.Let_def)`
to remove the let bindings.