impl: IMX serial driver for seL4.
This is a preliminary imlementation of the imx serial driver for
seL4(requires `utils_libs`).
Most of the Cogent code here in here does initialisation of driver.
The `putchar` and `getchar` operations are a little too simple for us
to implement it in Cogent. It can still be done, but I will hold it
off for the next iteration.

As mentioned in the README, the `Makefile` and the `include`
directives in `serial.ac` need to be updated to reflect your current
build system path for the the compiler to find `libutils` and
`platsupport`.