Add a simple iterator implementation
This adds an example that demonstrates how a simple iterator can be
constructed and used. It also shows how a program can be implemented
with a minimal amount of anti-quoted C.

The makefile here is also intended to demonstrate the bare minimum
needed to compile a full program from Cogent.