docs: minor tweaks to index and sec 1
[skip lemma]

But personally I don't like the landing page in its current structure.
The introduction section at the beginning to too long and
readers have to scroll down to the toc, which is quite annoying
and is not a good HCI experience. Also the description is
not at the right level that I would like.