minigent build: Stack build files for GHC 8.4, 8.6, 8.8.
Track the last release in the GHC 8.4, and 8.6 series by Stackage LTS
resolvers of appropriate vintage, which we pin in these files.

GHC 8.2 has a quirk related to Monoid/Semigroup that I couldn't quickly
hack around; so, for the moment, I'll consider it unsupported.

GHC 8.8 support is currently aspirational: Cabal 3.0, the MonadFail
changes to `base', and Stack's nascent support for GHC 8.8 mean this
isn't a supported configuration (yet).  Naturally, no LTS has been
cut (yet), so the resolver for 8.8 is today's nightly, and, once the
situation firms up, should be switched to an appropriate LTS.