compiler: remove lexicographic reordering of record fields.
As discussed in #333, we no longer want the Cogent compiler to
reorder the fields within a record. This commit redesigns the Row
module to remember the order specified by the Cogent programmer.

The Cogent compiler no longer identifies records with the same
fields written in different orders, and this commit adds a test
case to demonstrate this new behaviour.

I have removed unifiers corresponding to the "takeness" of an entry
within a row; there are currently no test cases which demonstrate
a need for this feature. The newly designed Row interface will make
it straightforward to add such a feature in the future should it
become necessary.

NB: Variants still reorder their components lexicographically. It
is not clear to me yet if we want to enforce no-reordering on
variants as well.