Changed value semantics embedding of word arrays
The new embedding of word arrays in the value semantics now contains the
type of the elements in the word array. The main benefit of this change
is that it removes duplication of abstractions for word array function,
due to different word sizes, which was previously thought to be
necessary. This then makes creating a tactic for proving scorres easier
since the same tactic will work even for the different word sizes
(assuming the correct lemmas and definitions are added the simpset).

Other changes were fixes to the definitions and proofs to reflect the
change of embedding of word arrays.