Changed update semantics embedding of word arrays
The embedding of word arrays has changed to include the type of the
element which is being stored. Currently, this change only supports
primitive number types , i.e. U8, U16, adn so on. Due to these changes,
the abstract typing defintion as well as the abstractions, were altered
to reflect these changes. Due to these changes, the proofs also had to
be updated. The major reason for this change was to allow for better
generalisation of definitions to reduce duplication of essentially the
same things only differing based on the element type.

One change that was made which was not caused by changing
the embedding was that the core abstractions for wordarray_fold_no_break
were changed to also take the the abstract function relation as an
argument. What this enables is that the same definition can now be used
for any function call level.