Replaced some lemmas with a more generalised version
The lemmas used to show that the monomorphising the abstractions of
wordarray_fold_no_break in the value semantics has been replaced with
one which does not assume the type of the accumalator and the observer
arguments to be of type 32-bit word and unit, respectively. This means
that this should be applicable for all wordarray_fold_no_break
functions.