Removed hard-coded abstraction relations from word array abstractions
This change removes the hard-coded xi abstraction relations from the
word array abstractions in both the update and value semantics. What
this means it that the same word array abstractions can be used for
different function call levels. It does this by taking the xi
abstraction relation as an argument. This should reduce the number of
definitions that would need to be generated.

One other changeis the complete generalisation for the value semantics
abstraction for wordarray_fold_no_break. This means that this
abstraction is now polymorphic over the accumulator and observer
arguments. The word array can also be of any standard word size which is
recognised by Cogent. Automation of the generation of the abstraction
for word_array_fold_no_break is no longer required for the value
semantics.