Added deterministic pointer constraint and lemma in the update semantics
The case where the uval typing constraint is true for the same value,
type, heap and abstract typing context but with different pointer sets,
should not occur. This is because then the value would not be equal to
itself since a value is made up of its pointer sets. In the case where
the pointer sets are different, this is clearly not the case. The
changes made prevent such a case occurring.

Other changes were just minor fixes to account for the above changes.