spi: fixed up prototype
Did not account for the fact that the size fields in the original
C code could be smaller or larger than the actual arrays given. So
the behaviour of the prototype and the C could be different behaviour.
So the size fields were included.

The 'rxbuf' field in the C version could in fact be NULL, although
this could be modelled as an empty list, using Maybe makes it more
obvious