wordarray: proved corres between shallow and value for mapAccum
Proved the correspondence between the shallow and the value semantics
embedding of mapAccum. In the process, it was found that the original
shallow embedding was incorrect when an invalid range was given. Note
that some of the other proofs may be broken but this is only due to
including mapAccum in the set of abstraction functions in the value
semantics, and the other correspondences have yet to be proved.