wordarray: proved mono-renaming and preservation for mapAccum
The mono-renaming lemma ensures that the polymorphic mapAccum does the
same thing as the monomorphic versions of mapAccum. The preservation
lemmas for the value and update semantic embeddings of mapAccum ensure
that the return value "vval type" and "uval type" correctly if the
arguments "vval type" and "uval type" correctly (plus a few extra
constraints on the environment and pointer sets). In addition, the frame
constraint is also satisfied.