wordarray WIP: cleaned up files
Shifted the update semantic embeddings of the word array functions and
the related proofs to "WordArray_Update.thy". Shifted the value semantic
embeddings of the word array functions and related proofs to
"WordArray_Value.thy". Shifted the xi abstractions to
"WordArray_UAbsFun.thy" and "WordArray_VAbsFun.thy" for the update and
value semantics respectively.