bilby wordarray: Changed some wordarray axioms to lemmas
Changed several axioms about word arrays, which are used in BilbyFS
proofs, to be lemmas instead. These axioms were for the word array
methods of length, get, map_no_break and for the abstraction functions.

The abstraction functions were changed so that they only worked for
lists and word arrays, whose elements were words. In addition,
constraints were added so that the length of the word array is bounded
by 2 ^ 32, due to that being the maximum length that a word array can be
represented with the current word array implementation.