wordarray: simplified deep embeddings
The deep embeddings for fold and map were simplified so that they no
longer rely on the value typing and the frame relation in their
embeddings. This reduces the size of of some of the core proofs but does
make some of the helper lemmas slightly longer in terms of lines of
proof. However, it is overall cleaner.