Added "crude" example demonstrating potential Dargent use case.
This example assumes we have an existing C struct we need to access
within Cogent but the struct has fields we do not want to reflect
in the Cogent specification. So we may create a space within the
Dargent layout for the "dont-care" fields but retain access to the
fields Cogent needs to know about. Then we can simply cast back
and forth between the C struct and the Dargent-generated
struct-of-byte-array.

[skip ci]