isa-parser: fix bugs introduced by earlier changes
the improved prettyprinter broke the generation of the deep embedding,
especially the table file for which parsing depends on the layout.
This is fixed by splitting the prettyprinter in a legacy version
and the improved version, where the improved version is only applied
to the shallow embedding in tuples form.