wordarray: Added C-update corres lemmas not dependant on the generated C
This commit includes correspondence lemmas between the C and update
semantics for the word array operations length, get and put. These
lemmas are designed to not rely on the generated C and instead rely on a
constant monadic C embedding of the implementation, which users then
equate with the generated C.

In addition, this commit provides correspondence for word arrays for
lemmas for the C generated from the Cogent code in this directory. These
are provided as validation that the generic update-C correspondence
lemmas for word arrays are reusable.