Shifted frame relation definition
The frame relation definition is shifted to before the definition of the
update semantics locale. This is done in preparation for changes to the
constraints of the update semantics locale, which will need to refer to
the frame relation.

This commit also contains minor changes to the proofs that broke due to
shifting the frame relation definition.