compiler proof: Fix final locale output
This commit modifies the AllRefine.hs file so that it correctly outputs
the final locale reflecting the changes made to the abs_typing in the
update semantics and abs_upd_val in the update-value semantics
correspondence.

This commit also fixes mistakes made in the previous commit in the
following files:
    * UpdateSemantics.thy: the assumption "f \<inter> u" in the
      abs_typing_frame was corrected to "r \<inter> u",
    * Correspondence.thy: the assumption
      "frame \<sigma> l \<sigma>' l" in the abs_upd_val_frame
      constraint was corrected to "frame \<sigma> l \<sigma>' l'".

This commit also changes the rule u_v_p_abs_ro in the Correspondence.thy
file, where previously it assumed
"abs_upd_val a a' n ts s r w \<sigma>", to now assume
"abs_upd_val a a' n ts s r {} \<sigma>". The rest of the changes to the
proofs reflect this change.