[Buffer TC] Add typechecking for buffer type and its arguments.
So far we check the first argument has an integral type, either
a literal or a variable bound to an integral type. It remains to
check the definition is a constant expression.

Regarding the second argument, we check that it is discardable.
Ideally, we need a more precise constraint that the argument is
unboxed.

I have had some test cases to demonstrate the checks are performed.