File clean up and added determinism and subset proofs
Two new files were added. These are "UpdateSemHelper.thy" and
"ValueSemHelper.thy". These two files contain helper lemmas and
definitions that are not specific to repeat.

Proofs showing that the update and value semantics are deterministic
if the abstract functions are deterministic were added.

Proofs showing that evaluation and preservation of abstract functions
in the update and value semantics are occurs for the same expressions
and arguments for an abstract function environment (xi) if it occurs
for a smaller abstract function environment were added.

Similarly, proofs for corres and correspondence of abstract functions
(for certain cases) were also added.