[Buffer] Added example of appending an entry onto the buffer.
We use the BTail type, representing the uninitialised portion of
the buffer, to append an entry in place. Now we must maintain the
length of the uninitialised segment in order to check there is
enough space for the new entry and update the length after append.

Abstract types are compiled to pointers so to support tracking the
additional length value we would have to allocate a structure
dynamically if we define BTail as an abstract type. Instead, we
make BTail concrete using abstract types for its fields to prevent
their modification from within Cogent; which would be unsafe in
general.