[Buffer/Memcpy] Added wip example demonstrating Dargent+Buffer approach.
I have yet to do anything interesting with the example but the
general idea is implemented: represent a contiguous block of objects
using a Buffer/Vector which all have a common header but vary in their
payload kind. Dargent is used to ensure all the objects in the block
are contiguous, in particular it forces all variant constructors to
start from the same bit offset.